+++ packages/aws-cdk-lib/core/lib/index.ts
@@ -39,7 +39,6 @@ export * from './duration';
 export * from './bitrate';
 export * from './expiration';
 export * from './size';
-export * from './stack-trace';
 export type { Element } from './deps';
 export * from './errors';
